apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "recall-api"
version = "0.1.0"
description = "Recall — AI-Native Work OS API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.3.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
full = [
    "weaviate-client>=4.9.0",
    "litellm>=1.50.0",
    "langgraph>=0.2.0",
    "temporalio>=1.7.0",
    "composio-core>=0.6.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
# B008: FastAPI ``Depends()`` defaults; E501: wrapped later; common try/except and typing-import noise
ignore = [
  "S101",
  "B008",
  "E501",
  "UP042",
  "S110",
  "S112",
  "S105",
  "B904",
  "TC001",
  "TC002",
  "S311",
  "S603",
  "E712",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
# Gradual typing: strict=False until annotations are expanded across routes/services
strict = false
ignore_missing_imports = true
warn_unused_configs = true
# Remaining noise is mostly ORM attribute timing and httpx/json Any — tighten per-module later
disable_error_code = ["attr-defined", "assignment", "arg-type", "operator", "union-attr", "list-item"]
